dsfdfds

+++ src/core/window.c
@@ -4347,7 +4347,7 @@ meta_window_adjust_opacity (MetaWindow   *window,
   }
 
   if (new_opacity != current_opacity)
-    meta_window_actor_set_opacity (actor, (guint8) new_opacity);
+    meta_window_actor_set_opacity (actor, new_opacity);
 }
 
 void
