+++ src/meta/meta-window-actor.h
@@ -66,7 +66,7 @@ gboolean           meta_window_actor_is_override_redirect (MetaWindowActor *self
 gboolean       meta_window_actor_showing_on_its_workspace (MetaWindowActor *self);
 gboolean       meta_window_actor_is_destroyed (MetaWindowActor *self);
 void           meta_window_actor_set_opacity  (MetaWindowActor *self,
-                                               guint8           opacity);
+                                               gint             opacity);
 guint8 meta_window_actor_get_opacity (MetaWindowActor *self);
 cairo_surface_t * meta_window_actor_get_image (MetaWindowActor       *self,
                                                cairo_rectangle_int_t *clip);
